requirements/base.txt
Django==5.2.6
pillow==11.3.0
django-cleanup==9.0.0
django-allauth==65.11.2
django-htmx==1.23.2
django-tailwind[reload]==4.2.0
django-widget-tweaks==1.5.0
resend==2.13.1
django-guardian==3.2.0